// frontend/src/models/propertyDefinitionsModelType.ts
export interface propertyDefinitionsModelType extends Logic {
    actionCreators: {
        loadPropertyDefinitions: (initial?: any) => {
            type: 'load property definitions (models.propertyDefinitionsModel)'
            payload: {
                initial: any
            }
        }
// ...
    }
// ...
    values: {
        propertyStorage: PropertyDefinitionStorage
        propertyStorageLoading: boolean
        loaded: boolean
        propertyDefinitions: PropertyDefinition[]
        transformedPropertyDefinitions: PropertySelectOption[]
        propertyNames: string[]
        numericalPropertyNames: PropertySelectOption[]
        describeProperty: (s: string) => string | null
        formatForDisplay: FormatForDisplayFunction
    }
// ...
}